requirements.txt
django
django-extensions
djangorestframework
pytest
pytest-django
requests
